requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.9
django-debug-toolbar==3.2.2
django-extensions==3.1.5
djangorestframework==3.12.4
drf-spectacular==0.21.0
et-xmlfile==1.1.0
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.2.1
mysqlclient==2.1.0
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.5.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
zipp==3.6.0
